refactor: api 로직 분리
