+++ src/Api/api.ts
@@ -1,43 +1,9 @@
-import type { ProductSummary, ThemeInfo, ThemeListItem } from '@/types/types';
+import type { ThemeInfo, ThemeListItem } from '@/types/types';
 import axios, { type AxiosResponse } from 'axios';
 
 export const api = axios.create({
   baseURL: import.meta.env.VITE_BASE_API_URL,
 });
-
-interface LoginResponse {
-  data: {
-    email: string;
-    name: string;
-    authToken: string;
-  };
-}
-
-interface ProductSummaryResponse {
-  data: ProductSummary;
-}
-
-export const postLogin = (email: string, password: string): Promise<AxiosResponse<LoginResponse>> =>
-  api.post<LoginResponse>('/api/login', { email, password });
-
-export const getProductSummary = async (productId: number): Promise<ProductSummary> => {
-  const { data } = await api.get<ProductSummaryResponse>(`/api/products/${productId}/summary`);
-  return data.data;
-};
-
-interface OrderBody {
-  productId: number;
-  message: string;
-  messageCardId: string;
-  ordererName: string;
-  receivers: {
-    name: string;
-    phoneNumber: string;
-    quantity: number;
-  }[];
-}
-
-export const postOrder = (body: OrderBody) => api.post('/api/order', body);
 
 api.interceptors.request.use((config) => {
   const raw = localStorage.getItem('auth');
@@ -96,5 +62,5 @@ export const getThemeProducts = async (themeId: number, cursor = 0, limit = 10) 
   const { data } = await api.get<ThemeProductResponse>(`/api/themes/${themeId}/products`, {
     params: { cursor, limit },
   });
-  return data.data; // { list, cursor, hasMoreList }
+  return data.data;
 };
